+++ internal/cache/vars.go
@@ -8,11 +8,13 @@ var (
 	SSProxiesCount            = 0
 	VmessProxiesCount         = 0
 	TrojanProxiesCount        = 0
+	VlessProxiesCount         = 0
 	UsefullProxiesCount       = 0
 	UsefullSSRProxiesCount    = 0
 	UsefullSSProxiesCount     = 0
 	UsefullVmessProxiesCount  = 0
 	UsefullTrojanProxiesCount = 0
+	UsefullVlessProxiesCount  = 0
 	LastCrawlTime             = "程序正在启动，请于3分钟后刷新页面"
 	IsSpeedTest               = "未开启"
 )